requirements.txt
python-dotenv==1.0.0
requests==2.31.0
gradio==4.19.2
streamlit==1.32.0 